2021 Day 17. Trick Shot.
The search space is constrained by exceeding the target range after 1
step.
So maximum X initial velocity to search is the maximum target X because
anything bigger would be out of the target area on the first step with
no hope of returning.

For positive Y velocity, the probe will eventually return to 0 height
with negative Y velocity of the same magnitude.